requirements.txt
cachetools
Faker==4.14.0
GitPython==3.1.41
Jinja2==3.1.3
Pillow==10.3.0
PyGithub==1.59.0
PyJWT==2.6.0
aiofiles==22.1.0
aiohttp==3.10.0
alembic>=1.7.5
bcrypt==4.0.1
croniter==1.3.7
cryptography==42.0.4
dask>=2022.2.0
datadog==0.44.0
freezegun==1.2.2
great-expectations==0.18.12
httpx==0.23.3
inflection==0.5.1
ipykernel==6.15.0
ipython==8.10.0
itsdangerous~=1.1.0
joblib>=1.1.0
jupyter-server==2.14.1
jupyter_client==7.4.4
ldap3==2.9.1
memory_profiler
newrelic==8.8.0
numpy>=1.22.0
pandas>=1.3.0
polars==1.29.0
protobuf>=4.25.0
psutil==5.9.8
pyairtable>=2.3.3
pyarrow>=14.0.1
python-dateutil==2.8.2
pytz>=2022.2.1
pyyaml~=6.0
redis~=5.0.1
requests~=2.31.0
ruamel.yaml==0.17.17
scikit-learn>=1.0
sentry-sdk==1.19.1
setuptools==70.0.0
simplejson
six>=1.15.0
sqlalchemy>=1.4.20, <2.0.0
sqlglot[rs]
terminado==0.17.1
thefuzz[speedup]==0.19.0
tornado==6.4.2
typer[all]==0.9.0
typing_extensions==4.10.0
watchdog==4.0.0
Werkzeug==3.0.3


# extras
astor>=0.8.1
aws-secretsmanager-caching==1.1.1.5
boto3==1.26.60
botocore==1.29.60
clickhouse-connect~=0.6.23
db-dtypes==1.0.5
duckdb==1.1.3
google-api-core~=2.15.0
google-api-python-client~=2.70.0
google-cloud-bigquery~=3.14.1
google-cloud-iam~=2.13.0
google-cloud-pubsub~=2.19.0
google-cloud-run~=0.10.1
google-cloud-storage~=2.5.0
gspread==5.7.2
kubernetes>=28.1.0
langchain==0.1.6; python_version >= '3.8'
langchain_community<0.0.20
mysql-connector-python~=8.4.0; python_version < '3.11'
mysql-connector-python~=9.0.0; python_version >= '3.11'
openai>=1.0.0
opentelemetry-api==1.22.0
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-instrumentation-tornado==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-sqlalchemy==0.43b0
oracledb==1.3.1; python_version < '3.12'
oracledb==2.4.1; python_version >= '3.12'
pinotdb==5.6.0
prometheus_client>=0.18.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyairtable==2.3.3
pydruid==0.6.5
pyodbc==4.0.35; python_version < '3.12'
pyodbc==5.0.1; python_version >= '3.12'
redshift-connector==2.0.915
lxml==4.9.4
snowflake-connector-python==3.7.1
sshtunnel==0.4.0
tables==3.7.0; python_version < '3.11'
tables==3.10.1; python_version >= '3.11'
trino~=0.326

# Azure
azure-eventhub==5.11.2
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-keyvault-certificates==4.6.0
azure-mgmt-containerinstance==10.1.0
azure-storage-blob==12.14.1

# Streaming
confluent-avro~=1.8.0
elasticsearch==8.15.1
influxdb_client==1.36.1
kafka-python==2.0.2
nats-py==2.6.0
nkeys~=0.2.0
opensearch-py==2.0.0
pika==1.3.1
pymongo==4.3.3
requests_aws4auth==1.1.2
stomp.py==8.1.0

# DBT
dbt-adapters==1.8.0
dbt-bigquery==1.8.3
dbt-clickhouse==1.8.5
dbt-core==1.8.7
dbt-duckdb==1.8.4
dbt-postgres==1.8.2
dbt-redshift==1.8.1
dbt-snowflake==1.8.4
dbt-spark==1.8.0
dbt-sqlserver==1.8.4
dbt-synapse==1.8.2
dbt-trino==1.8.4

async-timeout==4.0.3
attrs==22.2.0; python_version < '3.12'
attrs==24.2.0; python_version >= '3.12'
backoff
clickhouse_sqlalchemy
deltalake==0.20.2
facebook_business==22.0.2
gnupg==2.3.1
google-analytics-data==0.14.2; python_version < '3.11'
google-analytics-data==0.15.0; python_version >= '3.11'
google-cloud-bigquery-storage==2.16.2; python_version < '3.11'
google-cloud-bigquery-storage==2.26.0; python_version >= '3.11'
httpx[http2]
pendulum==2.1.0; python_version < '3.12'
pendulum==3.0.0; python_version >= '3.12'
pymongo_schema==0.4.1
python-magic==0.4.27
redshift-connector~=2.0.915
requests_mock==1.10.0
simple_salesforce~=1.12.5
singer_sdk~=0.34.1
stripe==5.5.0
terminaltables==3.1.10
twitter-ads==11.0.0
tzlocal==4.2
xlsx2csv==0.8.1
xmltodict==0.13.0
zenpy==2.0.25


#custom
minio
pyzipper
authlib
hvac
confluent_kafka==2.0.2
hydra-cli